Changes to module to make it more modular
