+++ README.md
@@ -3,6 +3,9 @@ The Voice Command Node Module is a project by Team Sonus at Southeastern Louisia
 This is a module that uses CMU PocketSphinx to recognize audio and respond to voice commands in an event based system.
 
 Justin LeJeune
+
 Nick Moran
+
 Tyler Becker
-Seth Champange
+
+Seth Champagne
