+++ index.js
@@ -1,0 +1,45 @@
+// Justin LeJeune - Sonus.js
+//
+// very basic command input for testing and demo
+//
+
+var sys     = require("sys")
+,   fs      = require('fs')
+,   sonus   = require('./sonus.js')
+,   stdin   = process.openStdin()
+,   word
+,   adder
+;
+
+console.log("\n\nPlease enter the the number for the operation you would like to run:");
+console.log(" 1.Run Sonus \n 2.Add a new word to the CMU dictionary \n 3.Add a new command \n 4.Exit");
+
+stdin.addListener("data", function(d) {
+    if (d.toString().trim() == 4) { 
+        die("this")
+
+    } else if (d.toString().trim() == 1) {
+        sonus();
+
+    } else if (d.toString().trim() == 2) {
+        console.log("\n\nEnter the word you want to add, followed by the pronunciation(separated by a pipe): ");
+        stdin.addListener("data", function(d) {
+            if (d.toString().trim()) {
+                var path = "wordList.js";
+
+                word = d.toString().trim();
+                word = word.split("|");
+
+                adder = "module.exports.push(['" + word[0] + "', '" + word[1] + "']);";
+
+                console.log(adder);
+                fs.appendFile("wordList.js", adder, function(err) {
+                    if (err) console.log(err);
+                });
+                
+            }
+        });
+    
+    }
+
+  });
