+++ lib/wordList.js
@@ -13,7 +13,7 @@ module.exports =    [
 
 ,   ["NEW-YORK",        "N UW Y AO R K"                 ]
 ,   ["NEW-YORK-CITY",   "N UW Y AO R K S IH T IY"       ]
-,   ["PARIS",           "P AE R IH S"                   ] 
+,   ["PARIS",           "P AE R IH S"                   ]
 ,   ["PARIS(2)",        "P EH R IH S"                   ]
 ,   ["SHANGHAI",        "SH AE NG HH AY"                ]
 ,   ["SAN-FRANCISCO",   "S AE N F R AE N S IH S K OW"   ]
@@ -53,4 +53,4 @@ module.exports =    [
 ,   ["A",               "AH"                            ]
 ,   ["A(2)",            "EY"                            ]
 ,   ["DICK",            "D IH K"                        ]
-];module.exports += [['this', 'that']];
+];
