+++ package.json
@@ -2,7 +2,7 @@
   "name": "sonus",
   "version": "1.0.0",
   "description": "Voice Command Node Module",
-  "main": "sonus.js",
+  "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
