--- voice/record.js~
@@ -1,30 +1,0 @@
-var fs          = require('fs')
-,   spawn       = require('child_process').spawn
-,   PassThrough = require('stream').PassThrough
-,   audio       = new PassThrough
-,   ps
-,   file
-;
-
-function start() { 
-    ps = spawn('sox', ['-t', 'waveaudio', '0', '-p']);
-    //ps.stdout.pipe(file);
-    //ps.stderr.pipe(info);
-};
-
-function stop() {
-    if(ps) {
-        ps.kill();
-        ps = null;
-    }
-};
-
-module.exports = function (fileName, seconds) {
-    file = fs.createWriteStream(fileName + '.wav', { encoding: 'binary' });
-    start();
-
-    setTimeout(function () {
-        stop();
-        file.end();
-    }, seconds * 1000);
-}
