--- voice/wordList.js
@@ -1,56 +1,0 @@
-// List of words to be loaded into recognizer
-module.exports =    [
-    ["ONE",             "W AH N"                        ]
-,   ["TWO",             "T UW"                          ]
-,   ["THREE",           "TH R IY"                       ]
-,   ["FOUR",            "F AO R"                        ]
-,   ["FIVE",            "F AY V"                        ]
-,   ["SIX",             "S IH K S"                      ]
-,   ["SEVEN",           "S EH V AH N"                   ]
-,   ["EIGHT",           "EY T"                          ]
-,   ["NINE",            "N AY N"                        ]
-,   ["ZERO",            "Z IH R OW"                     ]
-
-,   ["NEW-YORK",        "N UW Y AO R K"                 ]
-,   ["NEW-YORK-CITY",   "N UW Y AO R K S IH T IY"       ]
-,   ["PARIS",           "P AE R IH S"                   ] 
-,   ["PARIS(2)",        "P EH R IH S"                   ]
-,   ["SHANGHAI",        "SH AE NG HH AY"                ]
-,   ["SAN-FRANCISCO",   "S AE N F R AE N S IH S K OW"   ]
-,   ["LONDON",          "L AH N D AH N"                 ]
-,   ["BERLIN",          "B ER L IH N"                   ]
-
-,   ["SUCKS",           "S AH K S"                      ]
-,   ["ROCKS",           "R AA K S"                      ]
-,   ["IS",              "IH Z"                          ]
-,   ["NOT",             "N AA T"                        ]
-,   ["GOOD",            "G IH D"                        ]
-,   ["GOOD(2)",         "G UH D"                        ]
-,   ["GREAT",           "G R EY T"                      ]
-,   ["HELLO",           "HH AH L OW"                    ]
-,   ["WORLD",           "W ER L D"                      ]
-
-,   ["WINDOWS",         "W IH N D OW Z"                 ]
-,   ["LINUX",           "L IH N AH K S"                 ]
-,   ["UNIX",            "Y UW N IH K S"                 ]
-,   ["MAC",             "M AE K"                        ]
-,   ["O",               "OW"                            ]
-,   ["S",               "EH S"                          ]
-,   ["X",               "EH K S"                        ]
-
-,   ["AND",             "AE N D"                        ]
-,   ["AND(2)",          "AH N D"                        ]
-,   ["GO",              "G OW"                          ]
-,   ["FUCK",            "F AH K"                        ]
-,   ["YOURSELF",        "Y ER S EH L F"                 ]
-,   ["YOURSELF(2)",     "Y UH R S EH L F"               ]
-,   ["YOURSELF(3)",     "Y AO R S EH L F"               ]
-,   ["JUSTIN",          "JH AH S T AH N"                ]
-,   ["JUSTIN(2)",       "JH AH S T IH N"                ]
-,   ["AWESOME",         "AA S AH M"                     ]
-,   ["AWESOME(2)",      "AO S AH M"                     ]
-,   ["SUCK",            "S AH K"                        ]
-,   ["A",               "AH"                            ]
-,   ["A(2)",            "EY"                            ]
-,   ["DICK",            "D IH K"                        ]
-];
